The num of relevant sprites may differ from one scanline to the next
Before this commit, the same counter was used to register the number
of sprites which will be rendered in the next scanline, and to iterate
over the buffers of precalculated pixels in the current scanline. This
logic is wrong. If no sprites are relevant for the next scanline, then
no sprites will be rendered in the current one. Bad.