Fix bug in cycle-accurate sprite evaluation
Use auxiliary boolean variable to register if sprite 0 was hit.
Then, in the sprite fetch phase, overwrite global variable.
If the global variable, which is the one used at the pixel calculation
phase, is used instead, it may cancel what was calculated for the
scanline currently being rendered. Keep in mind that, in the
cycle-accurante implementation, sprite evaluation occurs during visible
cycles.