Decouple olc library from PPU, make PPU render library-agnostic
Changes were made across the project to decouple the core from whichever
library may be used for rendering game frames. The PPU will deposit a
new palette color code, together with its coordinates (scanline +
cycle), in a structure instantiated at the level above the core. Is is
the responsibility of the NES core's client to render the pixel.
This is only a first step on the way of turning this NES core into a
true black box.