Correct logical mistakes in pulse generator
Several aspects of the sweep unit and the pulse generator
in general were wrong. The concept of 'target period' was not
correctly mapped in this implementation. When the sweep unit is
clocked, the timer is overwritten with the LAST calculated
period (called targetPeriod), and a new targetPeriod is
prepared for when the sweep engine is clocked the next time.
Additionally, corrections were made in the conditions under which
the sweep unit mutes the pulse channel, and in the actions
that take place when the CPU updates the pulse channels'
timers.