Correct timing inaccuracy causing cracks in audio
SDL timestamps come in ms, but, given 60.1 fps, the period of
one frame is 16.638... ms. Prior to this commit, around .36 ms
were lost per each frame as a consequence.
This resulted in one frame not being calculated every 46 frames
approximately. Samples for the lost frame were not added to the
SDL queue, which resulted in unpleasant cracks.

A problem in the audio still needs addressing, though, and that is
the delay between actions displayed in video and the
corresponding audio.