Implement IRQ & NMI logic as objects inside the CPU
Alongside this change come several changes in the logic.
The NMI is implemented as an edge-triggered interrupt
(fired on falling edges). IRQs are kept as dependent
on the level.

Inside the frame counter engine in the APU, an IRQ may
be fired, but not cleared. IRQs are only cleared when reading
the status register or overwriting the irq inhibit flag.