Fix bug affecting vertical flip of 8x16 tiles
When vertically flipped, the top and bottom tiles
were flipped independently, and were still displayed
in their top and bottom positions. This is wrong.
When a 8x16 tile is flipped, the sprite's bottom
tile must be rendered in the top.

This bug was evident in Zelda, when secondary
characters using 8x16 tiles changed direction.